+++ apps/meteor/CHANGELOG.md
@@ -1,5 +1,58 @@
 # @rocket.chat/meteor
 
+## 6.5.0-rc.9
+
+### Patch Changes
+
+- Bump @rocket.chat/meteor version.
+  - @rocket.chat/core-typings@6.5.0-rc.9
+  - @rocket.chat/rest-typings@6.5.0-rc.9
+  - @rocket.chat/api-client@0.1.18-rc.2
+  - @rocket.chat/license@0.1.0-rc.9
+  - @rocket.chat/omnichannel-services@0.1.0-rc.9
+  - @rocket.chat/pdf-worker@0.0.24-rc.2
+  - @rocket.chat/presence@0.1.0-rc.9
+  - @rocket.chat/core-services@0.3.0-rc.9
+  - @rocket.chat/cron@0.0.20-rc.2
+  - @rocket.chat/gazzodown@3.0.0-rc.9
+  - @rocket.chat/model-typings@0.2.0-rc.9
+  - @rocket.chat/ui-contexts@3.0.0-rc.9
+  - @rocket.chat/server-cloud-communication@0.0.1
+  - @rocket.chat/fuselage-ui-kit@3.0.0-rc.9
+  - @rocket.chat/models@0.0.24-rc.2
+  - @rocket.chat/ui-theming@0.1.1-rc.0
+  - @rocket.chat/ui-client@3.0.0-rc.9
+  - @rocket.chat/ui-video-conf@3.0.0-rc.9
+  - @rocket.chat/web-ui-registration@3.0.0-rc.8
+  - @rocket.chat/instance-status@0.0.24-rc.2
+
+## 6.5.0-rc.8
+
+### Patch Changes
+
+- Bump @rocket.chat/meteor version.
+- 928177b602: Fix rocket.cat's app request message escaping
+  - @rocket.chat/core-typings@6.5.0-rc.8
+  - @rocket.chat/rest-typings@6.5.0-rc.8
+  - @rocket.chat/api-client@0.1.18-rc.1
+  - @rocket.chat/license@0.1.0-rc.8
+  - @rocket.chat/omnichannel-services@0.1.0-rc.8
+  - @rocket.chat/pdf-worker@0.0.24-rc.1
+  - @rocket.chat/presence@0.1.0-rc.8
+  - @rocket.chat/core-services@0.3.0-rc.8
+  - @rocket.chat/cron@0.0.20-rc.1
+  - @rocket.chat/gazzodown@3.0.0-rc.8
+  - @rocket.chat/model-typings@0.2.0-rc.8
+  - @rocket.chat/ui-contexts@3.0.0-rc.8
+  - @rocket.chat/server-cloud-communication@0.0.1
+  - @rocket.chat/fuselage-ui-kit@3.0.0-rc.8
+  - @rocket.chat/models@0.0.24-rc.1
+  - @rocket.chat/ui-theming@0.1.1-rc.0
+  - @rocket.chat/ui-client@3.0.0-rc.8
+  - @rocket.chat/ui-video-conf@3.0.0-rc.8
+  - @rocket.chat/web-ui-registration@3.0.0-rc.7
+  - @rocket.chat/instance-status@0.0.24-rc.1
+
 ## 6.5.0-rc.7
 
 ### Patch Changes