+++ ee/packages/api-client/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @rocket.chat/api-client
 
+## 0.1.18-rc.2
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.0-rc.9
+- @rocket.chat/rest-typings@6.5.0-rc.9
+
+## 0.1.18-rc.1
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.0-rc.8
+- @rocket.chat/rest-typings@6.5.0-rc.8
+
 ## 0.1.15-rc.7
 
 ### Patch Changes