+++ ee/packages/omnichannel-services/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @rocket.chat/omnichannel-services
 
+## 0.1.0-rc.9
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.0-rc.9
+- @rocket.chat/rest-typings@6.5.0-rc.9
+- @rocket.chat/pdf-worker@0.0.24-rc.2
+- @rocket.chat/core-services@0.3.0-rc.9
+- @rocket.chat/model-typings@0.2.0-rc.9
+- @rocket.chat/models@0.0.24-rc.2
+
+## 0.1.0-rc.8
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.0-rc.8
+- @rocket.chat/rest-typings@6.5.0-rc.8
+- @rocket.chat/pdf-worker@0.0.24-rc.1
+- @rocket.chat/core-services@0.3.0-rc.8
+- @rocket.chat/model-typings@0.2.0-rc.8
+- @rocket.chat/models@0.0.24-rc.1
+
 ## 0.1.0-rc.7
 
 ### Patch Changes