+++ packages/core-typings/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @rocket.chat/core-typings
 
+## 6.5.0-rc.9
+
+## 6.5.0-rc.8
+
 ## 6.5.0-rc.7
 
 ## 6.5.0-rc.6