+++ packages/livechat/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @rocket.chat/livechat Change Log
 
+## 1.14.9-rc.2
+
+### Patch Changes
+
+- @rocket.chat/gazzodown@3.0.0-rc.9
+
+## 1.14.9-rc.1
+
+### Patch Changes
+
+- @rocket.chat/gazzodown@3.0.0-rc.8
+
 ## 1.14.6-rc.7
 
 ### Patch Changes