+++ packages/ui-client/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @rocket.chat/ui-client
 
+## 3.0.0-rc.9
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@3.0.0-rc.9
+
+## 3.0.0-rc.8
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@3.0.0-rc.8
+
 ## 3.0.0-rc.7
 
 ### Patch Changes